fix(raster): address Copilot re-review on PR #122
- Raster.h: DrawItem::cull is now CullMode (was raw u8) — type-safe,
  self-documenting call sites; Raster.cpp casts to u8 at the setup_triangle
  boundary.
- TriSetup.cpp: switch the cull branches so Back is the default/fallback —
  an unexpected cull_mode value now culls (Back) instead of silently going
  two-sided (None). Fix two stale/incorrect comments: the w<=0 reject is the
  eye-plane test (near-plane straddlers are split upstream now), and the
  front-facing convention is CCW *after* the Y-flip (positive screen area),
  not CCW-in-NDC.
- MeshWinding.h: fix_winding takes vertex_count and skips index triples that
  reference an out-of-range vertex (no OOB reads), matching the rasterizer's
  own guard; all sample call sites updated.
- tests: use CullMode::{Back,Front,None} instead of magic 0/1/2 so the tests
  track the enum if its values ever change.

+++ engine/render/raster/Raster.cpp
@@ -307,7 +307,7 @@ void Rasterizer::end_frame() {
                                                fs.view.target.width,
                                                fs.view.target.height,
                                                tris[produced],
-                                               d.cull);
+                                               static_cast<u8>(d.cull));
                 // Advance only on success so cmd.tri_count counts valid tris and
                 // the binner never iterates culled / degenerate setups.
                 if (ok)
